Migrate message unread method to new structure
