--- app/message-mark-as-unread/server/index.js
@@ -1,1 +1,0 @@
-import './unreadMessages';
