--- app/message-mark-as-unread/server/logger.js
@@ -1,9 +1,0 @@
-import { Logger } from '../../logger';
-
-const logger = new Logger('MessageMarkAsUnread', {
-	sections: {
-		connection: 'Connection',
-		events: 'Events',
-	},
-});
-export default logger;
