+++ server/api/methods/index.js
@@ -1,0 +1,7 @@
+import './2fa/checkCodesRemaining';
+import './2fa/disable';
+import './2fa/enable';
+import './2fa/regenerateCodes';
+import './2fa/validateTempToken';
+import './mentions/getUserMentionsByChannel';
+import './messages/unreadMessages';
