+++ server/api/methods/messages/unreadMessages.js
@@ -1,0 +1,7 @@
+import { Meteor } from 'meteor/meteor';
+
+import { unreadMessages } from '../../../services/messages/functions/unreadMessages';
+
+Meteor.methods({
+	unreadMessages,
+});
