+++ server/importPackages.js
@@ -40,7 +40,6 @@ import '../app/token-login/server';
 import '../app/mailer';
 import '../app/mapview/server';
 import '../app/markdown/server';
-import '../app/message-mark-as-unread/server';
 import '../app/message-pin/server';
 import '../app/message-snippet/server';
 import '../app/message-star/server';
@@ -88,6 +87,7 @@ import '../app/livechat/server';
 import '../app/custom/server';
 import '../app/authentication/server';
 
+import './api/methods';
 import './services/2fa';
 import './api/v1';
 import './services/messages/slashcommands';
