+++ server/services/2fa/index.js
@@ -1,8 +1,3 @@
 import './MethodInvocationOverride';
 import '../../settings/settingfiles/2fa.settings';
-import '../../api/methods/2fa/checkCodesRemaining';
-import '../../api/methods/2fa/disable';
-import '../../api/methods/2fa/enable';
-import '../../api/methods/2fa/regenerateCodes';
-import '../../api/methods/2fa/validateTempToken';
 import './loginHandler';
